// hub/frontend/lib/api/generated/model/skillAuditHistoryResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.129
 */
import type { SkillAuditHistoryEntryRead } from './skillAuditHistoryEntryRead';

export interface SkillAuditHistoryResponse {
  data: SkillAuditHistoryEntryRead[];
  id: string;
  slug: string;
  source: string;
}